Add env vars for tf configuration files
