+++ renovate.json
@@ -1,0 +1,4 @@
+{
+    "$schema": "https://docs.renovatebot.com/renovate-schema.json",
+    "extends": ["github>app-sre/shared-pipelines//renovate/default.json"]
+  }
